Incorporate changes & Fix previous changes
